Refactors MG webhook endpoint to webhook.ts
